+++ types/resources/CreditNote.d.ts
@@ -327,6 +327,8 @@ Ireland** . The first two characters of the VAT number in such a case is &#x60;X
     billing_address?:CreditNote.BillingAddress;
     
     site_details_at_creation?:CreditNote.SiteDetailsAtCreation;
+
+    tax_origin?:CreditNote.TaxOrigin;
   }
   export namespace CreditNote {
     export class CreditNoteResource {  
@@ -1784,6 +1786,11 @@ If you have enabled [EU VAT](https://www.chargebee.com/docs/eu-vat.html) in 2021
       timezone?:string;
        
       organization_address?:object;
+    }
+    export interface TaxOrigin {  
+      country?:string;
+       
+      registration_number?:string;
     }
   }
 }